- Refector Lifecycle Status - Add interested as status
