+++ src/main/java/de/samply/bbmri/negotiator/control/owner/OwnerQueriesDetailBean.java
@@ -53,9 +53,9 @@ import de.samply.bbmri.negotiator.model.OfferPersonDTO;
 import de.samply.bbmri.negotiator.model.OwnerQueryStatsDTO;
 import de.samply.bbmri.negotiator.rest.RestApplication;
 import de.samply.bbmri.negotiator.rest.dto.QueryDTO;
-import de.samply.bbmri.negotiator.util.CollectionLifeCycleStatus;
+import eu.bbmri.eric.csit.service.negotiator.lifecycle.CollectionLifeCycleStatus;
 import de.samply.bbmri.negotiator.util.DataCache;
-import de.samply.bbmri.negotiator.util.RequestLifeCycleStatus;
+import eu.bbmri.eric.csit.service.negotiator.lifecycle.RequestLifeCycleStatus;
 import org.jooq.Record;
 import org.jooq.Result;
 import org.slf4j.Logger;
