--- src/main/java/eu/bbmri/eric/csit/service/negotiator/lifeCycle/util/LifeCycleStatusUtil.java
@@ -1,4 +1,0 @@
-package eu.bbmri.eric.csit.service.negotiator.lifeCycle.util;
-
-public class LifeCycleStatusUtil {
-}
