+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/lifecycle/requeststatus/RequestStatusAccepptCondition.java
@@ -1,4 +1,4 @@
-package eu.bbmri.eric.csit.service.negotiator.lifeCycle.requestStatus;
+package eu.bbmri.eric.csit.service.negotiator.lifecycle.requeststatus;
 
 import de.samply.bbmri.negotiator.model.CollectionRequestStatusDTO;
 
@@ -6,17 +6,19 @@ import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 
-public class RequestStatusEndOfProject implements RequestStatus {
+public class RequestStatusAccepptCondition implements RequestStatus {
 
     private String status = null;
-    private String statusType = "endOfProject";
-    private String statusText = "End of Project announced.";
+    private String statusType = "accepptConditions";
+    private String statusText = "Acceppt Access Condition indicated for collection and select for further work.";
     private Date statusDate = null;
-    private List allowedNextStatus = Arrays.asList("not_interrested", "offer");
-    private List allowedNextStatusBiobanker = Arrays.asList("abandoned.not_interrested");
-    private List allowedNextStatusResearcher = Arrays.asList("notselected.notselected", "dataReturnOffer.offer");
+    private List allowedNextStatus = Arrays.asList("not_interested", "signed");
 
-    public RequestStatusEndOfProject(CollectionRequestStatusDTO collectionRequestStatusDTO) {
+    private List allowedNextStatusBiobanker = Arrays.asList("notselected.notselected", "mtaSigned.signed", "abandoned.not_interested");
+
+    private List allowedNextStatusResearcher = Arrays.asList("notselected.watingForResponse", "abandoned.not_interested");
+
+    public RequestStatusAccepptCondition(CollectionRequestStatusDTO collectionRequestStatusDTO) {
         statusDate = collectionRequestStatusDTO.getStatusDate();
         status = collectionRequestStatusDTO.getStatus();
     }
