+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/lifecycle/requeststatus/RequestStatusContact.java
@@ -1,24 +1,27 @@
-package eu.bbmri.eric.csit.service.negotiator.lifeCycle.requestStatus;
+package eu.bbmri.eric.csit.service.negotiator.lifecycle.requeststatus;
 
 import de.samply.bbmri.negotiator.model.CollectionRequestStatusDTO;
+import eu.bbmri.eric.csit.service.negotiator.lifecycle.util.LifeCycleStatusUtilNextStatus;
 
 import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 
-public class RequestStatusMTASigned implements RequestStatus {
+public class RequestStatusContact implements RequestStatus {
 
     private String status = null;
-    private String statusType = "mtaSigned";
-    private String statusText = "MTA and Paperwork for shipping signed.";
+    private String statusType = "contact";
+    private String statusText = "Collection representatives not contacted yet.";
     private Date statusDate = null;
-    private List allowedNextStatus = Arrays.asList("not_interrested", "shipped");
+    private List allowedNextStatus = LifeCycleStatusUtilNextStatus.getAllowedNextStatus(this.getClass().getName());
 
-    private List allowedNextStatusBiobanker = Arrays.asList("notselected.notselected", "shippedSamples.shipped", "abandoned.not_interrested");
+    private List allowedNextStatusBiobanker = Arrays.asList("notselected.notselected", "availability.sample_data_available_accessible",
+            "availability.sample_data_available_not_accessible", "availability.sample_data_not_available_collecatable",
+            "availability.sample_data_not_available", "abandoned.not_interested");
 
     private List allowedNextStatusResearcher = Arrays.asList("notselected.watingForResponse");
 
-    public RequestStatusMTASigned(CollectionRequestStatusDTO collectionRequestStatusDTO) {
+    public RequestStatusContact(CollectionRequestStatusDTO collectionRequestStatusDTO) {
         statusDate = collectionRequestStatusDTO.getStatusDate();
         status = collectionRequestStatusDTO.getStatus();
     }
@@ -40,7 +43,14 @@ public class RequestStatusMTASigned implements RequestStatus {
 
     @Override
     public String getStatusText() {
-        return statusText;
+        if(status == null) {
+            return statusText;
+        } else if (status.equals("contacted")) {
+            return "Collection representatives contacted.";
+        } else if (status.equals("notreachable")) {
+            return "Collection representatives not reachable, BBMRI-ERIC has been informed.";
+        }
+        return "ERROR";
     }
 
     @Override
