+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/lifecycle/requeststatus/RequestStatusCreate.java
@@ -1,19 +1,21 @@
-package eu.bbmri.eric.csit.service.negotiator.lifeCycle.requestStatus;
+package eu.bbmri.eric.csit.service.negotiator.lifecycle.requeststatus;
 
 import de.samply.bbmri.negotiator.model.RequestStatusDTO;
+import eu.bbmri.eric.csit.service.negotiator.lifecycle.util.LifeCycleStatusUtilNextStatus;
 
 import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 
-public class RequestStatusFinish implements RequestStatus {
+public class RequestStatusCreate implements RequestStatus {
 
-    private String statusType = "finish";
-    private String statusText = "Request finished";
+    private String status = "created";
+    private String statusType = "created";
+    private String statusText = "Request created";
     private Date statusDate = null;
-    private List allowedNextStatus = Arrays.asList("review");
+    private List allowedNextStatus = LifeCycleStatusUtilNextStatus.getAllowedNextStatus(this.getClass().getName());
 
-    public RequestStatusFinish(RequestStatusDTO requestStatus) {
+    public RequestStatusCreate(RequestStatusDTO requestStatus) {
         statusDate = requestStatus.getStatusDate();
     }
 
@@ -24,7 +26,7 @@ public class RequestStatusFinish implements RequestStatus {
 
     @Override
     public String getStatus() {
-        return "finished";
+        return status;
     }
 
     @Override
@@ -44,7 +46,7 @@ public class RequestStatusFinish implements RequestStatus {
 
     @Override
     public List getAllowedNextStatus() {
-        return null;
+        return allowedNextStatus;
     }
 
     @Override
@@ -59,6 +61,6 @@ public class RequestStatusFinish implements RequestStatus {
 
     @Override
     public String getTableRow() {
-        return "<tr><td>" + statusDate + "</td><td>finish</td><td></td><td></tr>";
+        return "<tr><td>" + statusDate + "</td><td>created</td><td></td><td></tr>";
     }
 }
