+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/lifecycle/requeststatus/RequestStatusDataReturnOffer.java
@@ -1,29 +1,43 @@
-package eu.bbmri.eric.csit.service.negotiator.lifeCycle.requestStatus;
+package eu.bbmri.eric.csit.service.negotiator.lifecycle.requeststatus;
 
 import de.samply.bbmri.negotiator.model.CollectionRequestStatusDTO;
 import org.jooq.tools.json.JSONObject;
 import org.jooq.tools.json.JSONParser;
 import org.jooq.tools.json.ParseException;
 
-import java.util.Arrays;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-public class RequestStatusShippedSampesData implements RequestStatus {
+public class RequestStatusDataReturnOffer implements RequestStatus {
 
     private String status = null;
-    private String statusType = "shippedSamples";
-    private String statusText = "Shipped Samples/Data.";
+    private String statusType = "dataReturnOffer";
+    private String statusText = "Offer for data return.";
     private Date statusDate = null;
-    private List allowedNextStatus = Arrays.asList("not_interrested", "received");
-    private List allowedNextStatusBiobanker = Arrays.asList("abandoned.not_interrested");
-    private List allowedNextStatusResearcher = Arrays.asList("notselected.notselected", "receivedSamples.received");
+    private List<String> allowedNextStatus = new ArrayList<String>();
+    private List<String> allowedNextStatusBiobanker = new ArrayList<String>();
+    private List<String> allowedNextStatusResearcher = new ArrayList<String>();
 
-    public RequestStatusShippedSampesData(CollectionRequestStatusDTO collectionRequestStatusDTO) {
+    public RequestStatusDataReturnOffer(CollectionRequestStatusDTO collectionRequestStatusDTO) {
         statusDate = collectionRequestStatusDTO.getStatusDate();
         status = collectionRequestStatusDTO.getStatus();
-        if(status.equals("shipped")) {
-            statusText = "Shipped Number: " + getStatusTextFromJson(collectionRequestStatusDTO.getStatusJson(), "shippedNumber");
+        if(status == null) {
+            allowedNextStatus.add("not_interested");
+            allowedNextStatusBiobanker.add("notselected.notselected");
+            allowedNextStatusBiobanker.add("abandoned.not_interested");
+            allowedNextStatusResearcher.add("notselected.watingForResponse");
+        }
+        if(status.equals("offer")) {
+            statusText = "Data return offer: " + getStatusTextFromJson(collectionRequestStatusDTO.getStatusJson(), "offer");
+            allowedNextStatus.add("not_interested");
+            allowedNextStatus.add("accepted");
+            allowedNextStatus.add("rejected");
+            allowedNextStatusBiobanker.add("notselected.notselected");
+            allowedNextStatusBiobanker.add("dataReturnOffer.accepted");
+            allowedNextStatusBiobanker.add("dataReturnOffer.rejected");
+            allowedNextStatusBiobanker.add("abandoned.not_interested");
+            allowedNextStatusResearcher.add("notselected.watingForResponse");
         }
     }
