+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/lifecycle/requeststatus/RequestStatusEndOfProject.java
@@ -1,4 +1,4 @@
-package eu.bbmri.eric.csit.service.negotiator.lifeCycle.requestStatus;
+package eu.bbmri.eric.csit.service.negotiator.lifecycle.requeststatus;
 
 import de.samply.bbmri.negotiator.model.CollectionRequestStatusDTO;
 
@@ -6,17 +6,17 @@ import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 
-public class RequestStatusReceivedSamplesData implements RequestStatus {
+public class RequestStatusEndOfProject implements RequestStatus {
 
     private String status = null;
-    private String statusType = "receivedSamples";
-    private String statusText = "Samples/Data received.";
+    private String statusType = "endOfProject";
+    private String statusText = "End of Project announced.";
     private Date statusDate = null;
-    private List allowedNextStatus = Arrays.asList("not_interrested", "end");
-    private List allowedNextStatusBiobanker = Arrays.asList("abandoned.not_interrested");
-    private List allowedNextStatusResearcher = Arrays.asList("notselected.notselected", "endOfProject.end");
+    private List allowedNextStatus = Arrays.asList("not_interested", "offer");
+    private List allowedNextStatusBiobanker = Arrays.asList("abandoned.not_interested");
+    private List allowedNextStatusResearcher = Arrays.asList("notselected.notselected", "dataReturnOffer.offer");
 
-    public RequestStatusReceivedSamplesData(CollectionRequestStatusDTO collectionRequestStatusDTO) {
+    public RequestStatusEndOfProject(CollectionRequestStatusDTO collectionRequestStatusDTO) {
         statusDate = collectionRequestStatusDTO.getStatusDate();
         status = collectionRequestStatusDTO.getStatus();
     }
