+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/lifecycle/requeststatus/RequestStatusReceivedSamplesData.java
@@ -1,4 +1,4 @@
-package eu.bbmri.eric.csit.service.negotiator.lifeCycle.requestStatus;
+package eu.bbmri.eric.csit.service.negotiator.lifecycle.requeststatus;
 
 import de.samply.bbmri.negotiator.model.CollectionRequestStatusDTO;
 
@@ -6,19 +6,17 @@ import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 
-public class RequestStatusAccepptCondition implements RequestStatus {
+public class RequestStatusReceivedSamplesData implements RequestStatus {
 
     private String status = null;
-    private String statusType = "accepptConditions";
-    private String statusText = "Acceppt Access Condition indicated for collection and select for further work.";
+    private String statusType = "receivedSamples";
+    private String statusText = "Samples/Data received.";
     private Date statusDate = null;
-    private List allowedNextStatus = Arrays.asList("not_interrested", "signed");
+    private List allowedNextStatus = Arrays.asList("not_interested", "end");
+    private List allowedNextStatusBiobanker = Arrays.asList("abandoned.not_interested");
+    private List allowedNextStatusResearcher = Arrays.asList("notselected.notselected", "endOfProject.end");
 
-    private List allowedNextStatusBiobanker = Arrays.asList("notselected.notselected", "mtaSigned.signed", "abandoned.not_interrested");
-
-    private List allowedNextStatusResearcher = Arrays.asList("notselected.watingForResponse", "abandoned.not_interrested");
-
-    public RequestStatusAccepptCondition(CollectionRequestStatusDTO collectionRequestStatusDTO) {
+    public RequestStatusReceivedSamplesData(CollectionRequestStatusDTO collectionRequestStatusDTO) {
         statusDate = collectionRequestStatusDTO.getStatusDate();
         status = collectionRequestStatusDTO.getStatus();
     }
