+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/lifecycle/util/LifeCycleStatusType.java
@@ -1,0 +1,11 @@
+package eu.bbmri.eric.csit.service.negotiator.lifecycle.util;
+
+public class LifeCycleStatusType {
+    public static final String NOT_INTERESTED = "not_interested";
+    public static final String INTERESTED = "interested";
+    public static final String UNDER_REVIEW = "under_review";
+    public static final String ABANDONED = "abandoned";
+    public static final String WAITING_START = "waitingstart";
+    public static final String APPROVED = "approved";
+    public static final String REJECTED = "rejected";
+}
