+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/notification/NotificationService.java
@@ -1,11 +1,7 @@
 package eu.bbmri.eric.csit.service.negotiator.notification;
 
-import de.samply.bbmri.negotiator.Config;
-import de.samply.bbmri.negotiator.ConfigFactory;
-import de.samply.bbmri.negotiator.db.util.DbUtil;
 import de.samply.bbmri.negotiator.jooq.tables.records.NotificationRecord;
 import eu.bbmri.eric.csit.service.negotiator.database.DatabaseUtil;
-import eu.bbmri.eric.csit.service.negotiator.database.DatabaseUtilNotification;
 import eu.bbmri.eric.csit.service.negotiator.notification.types.*;
 import eu.bbmri.eric.csit.service.negotiator.notification.util.NotificationType;
 import org.slf4j.Logger;
@@ -25,7 +21,7 @@ public class NotificationService {
     }
 
     public static void sendNotification(Integer notificationType, Integer requestId, Integer commentId, Integer personId, Map<String, String> parameters) {
-        logger.info("23afa6c4695a-NotificationService: %d requestID: %d commentID: %d", notificationType, requestId, commentId);
+        logger.info("23afa6c4695a-NotificationService: {} requestID: {} commentID: {}", notificationType, requestId, commentId);
 
         NotificationRecord notificationRecord = createNotificationEntry(notificationType, requestId, commentId, personId);
         if(notificationRecord == null) {
@@ -67,8 +63,7 @@ public class NotificationService {
 
     private static NotificationRecord createNotificationEntry(Integer notificationType, Integer requestId, Integer commentId, Integer personId) {
         try {
-            NotificationRecord notificationRecord = databaseUtil.getDatabaseUtilNotification().addNotificationEntry(notificationType, requestId, commentId, personId);
-            return notificationRecord;
+            return databaseUtil.getDatabaseUtilNotification().addNotificationEntry(notificationType, requestId, commentId, personId);
         } catch (Exception ex) {
             logger.error("23afa6c4695a-NotificationService ERROR-NG-0000009: Create Notification Entry.");
             logger.error("context", ex);
