+++ src/main/resources/global.properties
@@ -241,7 +241,7 @@ endOfProject.end=Announce End of Project
 dataReturnOffer.offer=Offer for Data Return
 dataReturnOffer.accepted=Accept data return offer
 dataReturnOffer.rejected=Reject data return offer
-abandoned.not_interrested=Step away
+abandoned.not_interested=Step away
 notselected.notselected=-- Select Next Status --
 notselected.watingForResponse=Waiting for representative to responde
 notselected.watingForResponseFromResearcher=Waiting for representative from researcher
