+++ src/test/java/de/samply/bbmri/negotiator/util/requestStatus/RequestStatusCreateTest.java
@@ -1,7 +1,7 @@
 package de.samply.bbmri.negotiator.util.requestStatus;
 
 import de.samply.bbmri.negotiator.model.RequestStatusDTO;
-import eu.bbmri.eric.csit.service.negotiator.lifeCycle.requestStatus.RequestStatusCreate;
+import eu.bbmri.eric.csit.service.negotiator.lifecycle.requeststatus.RequestStatusCreate;
 import org.junit.Before;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -80,7 +80,7 @@ public class RequestStatusCreateTest {
     @Test
     @DisplayName("Test check all allowed next status.")
     void testAllowedNextStatusList() {
-        List allowedNextStatus = Arrays.asList("under_review");
+        List allowedNextStatus = Arrays.asList("under_review", "abandoned");
         assertEquals(allowedNextStatus, requestStatusCreate.getAllowedNextStatus());
     }
